+++ pkgs/python-packages/budget-report/default.nix
@@ -1,16 +1,12 @@
-{ callPackage, pytestCheckHook, buildPythonPackage, gspread,
-# fuzzywuzzy,
-easy-google-auth, gmail-parser }:
+{ callPackage, pytestCheckHook, buildPythonPackage, gspread, fuzzywuzzy
+, easy-google-auth, gmail-parser }:
 callPackage ../pythonPkgFromScript.nix {
   pname = "budget_report";
   version = "1.0.0";
   description = "Generate a budget report.";
-  # script-file = ./report.py;
-  script-file = ./new.report.py;
+  script-file = ./report.py;
   inherit pytestCheckHook buildPythonPackage;
-  propagatedBuildInputs = [ gspread 
-    # fuzzywuzzy 
-    easy-google-auth gmail-parser ];
+  propagatedBuildInputs = [ gspread fuzzywuzzy easy-google-auth gmail-parser ];
   checkPkgs = [ ];
   longDescription = ''
     ```
